-Added filtering -Added sorting -Added UI customization -Added lifetime clear fetching -Added caching system

+++ src-tauri/src/pollers/overlay.rs
@@ -33,7 +33,7 @@ impl Poller {
         let notification_state = unsafe { SHQueryUserNotificationState() };
 
         match notification_state {
-            Ok(n) if n.0 == 3 => return PollResult::Closed, // If in DX exclusive fullscreen mode
+            Ok(n) if n.0 == 3 => return PollResult::Closed,
             _ => (),
         }
 
@@ -132,8 +132,8 @@ pub async fn overlay_poller(handle: AppHandle) {
 
                 overlay
                     .set_size(PhysicalSize {
-                        width: dims.right - dims.left,
-                        height: dims.bottom - dims.top,
+                        width: (dims.right - dims.left) as u32,
+                        height: (dims.bottom - dims.top) as u32,
                     })
                     .unwrap();
